Use a Fullmatrix<float> instead one of doubles, to avoid *large* amounts of memory being eaten up! The memory consumption is still roughly quadratic in the number of degrees of freedom, but at least reduced to half of the previous value.
git-svn-id: https://svn.dealii.org/trunk@3505 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/include/dofs/dof_tools.h
@@ -886,7 +886,7 @@ class DoFTools
 			       const InterGridMap<DoFHandler,dim> &coarse_to_fine_grid_map,
 			       const vector<Vector<double> >      &parameter_dofs,
 			       const vector<int>                  &weight_mapping,
-			       FullMatrix<double>                 &weights);
+			       FullMatrix<float>                  &weights);
 
 				     /**
 				      * This is a function that is
@@ -905,7 +905,7 @@ class DoFTools
 				 const InterGridMap<DoFHandler,dim> &coarse_to_fine_grid_map,
 				 const vector<Vector<double> >      &parameter_dofs,
 				 const vector<int>                  &weight_mapping,
-				 FullMatrix<double>                 &weights,
+				 FullMatrix<float>                  &weights,
 				 const typename DoFHandler<dim>::active_cell_iterator &begin,
 				 const typename DoFHandler<dim>::active_cell_iterator &end);
 };
